requirements.txt
# Windows Memory Forensics & Malware Detection Tool

# Core Dependencies
yara-python>=4.3.1
psutil>=5.9.8
pywin32>=306
networkx>=3.0

# Machine Learning
numpy>=1.26.0
scikit-learn>=1.3.2
pandas>=2.1.0
joblib>=1.3.2
openpyxl>=3.1.0

# UI Framework
PyQt5>=5.15.10

# Reporting
reportlab>=4.0.7
Pillow>=10.1.0

# Security & Hashing
pycryptodome>=3.19.0

# Additional required package
pefile>=2023.2.7

# Utilities
colorama>=0.4.6
tqdm>=4.66.1